Separating plots to new file. DRY code. Need to work on slider filter with plots
