+++ src/dateFunctions.js
@@ -141,4 +141,34 @@ function convertDateFormat(input) {
   return trialConvert;
 }
 
-module.exports = {convertDateFormat, getDayNumber, getWeekDay, getMonthName};
+function last2000Days() {
+  result = [];
+  for (var i = 0; i < 2000; i++) {
+    var d = new Date();
+    d.setDate(d.getDate() - i);
+    result.push(formatDate(d));
+  }
+}
+
+function formatDate(date) {
+  var dd = date.getDate();
+  var mm = date.getMonth() + 1;
+  var yyyy = date.getFullYear();
+  if (dd < 10) {
+    dd = "0" + dd;
+  }
+  if (mm < 10) {
+    mm = "0" + mm;
+  }
+  date = mm + "/" + dd + "/" + yyyy;
+  return date;
+}
+
+module.exports = {
+  convertDateFormat,
+  getDayNumber,
+  getWeekDay,
+  getMonthName,
+  last2000Days,
+  formatDate,
+};
