+++ src/index.html
@@ -5,10 +5,6 @@
       rel="stylesheet prefetch"
       href="https://cdnjs.cloudflare.com/ajax/libs/rangeslider.js/2.3.0/rangeslider.min.css"
     />
-    <link
-    href="https://fonts.googleapis.com/css?family=Quicksand:300,400"
-    rel="stylesheet prefetch"
-  />
     <style class="cp-pen-styles"></style>
     <title>Cryptophan: Cryptocurrency Data</title>
     <!-- <meta name="viewport" content="initial-scale=1.0;" /> -->
