+++ src/index.js
@@ -1,13 +1,33 @@
-require('normalize.css/normalize.css');
-require ('./styles.css');
-const {convertDateFormat, getDayNumber, getWeekDay, getMonthName} = require ('./dateFunctions.js');
+require("normalize.css/normalize.css");
+require("./styles.css");
+const {
+  convertDateFormat,
+  getDayNumber,
+  getWeekDay,
+  getMonthName,
+  last2000Days,
+  formatDate,
+} = require("./dateFunctions.js");
+const {
+  plotPriceHistory,
+  plotWeekdayChange,
+  plotATHHistory,
+  plotHourlyData,
+} = require("./plots.js");
 
-ethereumAddresses = [];
-samePrice = [];
-allTimeHighPrice = 0;
-menuItem = 0;
-filterDays = 365;
-exchange = "CCCAGG";
+let samePrice = [];
+let allTimeHighPrice = 0;
+let navMenuPosition = 0;
+let exchange = "CCCAGG";
+let coinNumerator = "";
+let coinDenominator = "";
+const options = {
+  method: "GET",
+  headers: {
+    Accept: "application/json",
+    "Content-Type": "application/json;charset=UTF-8",
+  },
+};
 $(document).ready(function () {
   var localDate = new Date();
   timeOffset = localDate.getTimezoneOffset() / 60;
@@ -19,29 +39,16 @@ $(document).ready(function () {
   $(".date-info").html(weekDay + ", " + month + " " + dayLong + " ");
 });
 
-function getHistoricalPriceData(
-  coinNumerator,
-  coinDenominator,
-  unixEntered,
-  divide
-) {
+function getHistoricalPriceData(divide) {
+  console.log("getHistoricalPriceData function");
   const url =
     "https://min-api.cryptocompare.com/data/histoday?fsym=" +
     coinNumerator +
     "&tsym=" +
     coinDenominator +
-    "&limit=2000&aggregate=1&toTs=" +
-    unixEntered +
-    "&e=" +
+    "&limit=2000&aggregate=1&toTs=-2211638400&e=" +
     exchange;
   const urlEntered = divide ? url + "&tryConversion=false" : url;
-  const options = {
-    method: "GET",
-    headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json;charset=UTF-8",
-    },
-  };
 
   fetch(urlEntered, options)
     .then((response) => response.json())
@@ -55,9 +62,13 @@ function getHistoricalPriceData(
       var historicalDate = [];
       var historicalChange = [];
       var historicalWeekday = [];
-
+      allTimeHighRatio = [];
+      allTimeHighRatioDates = [];
+      allTimeHighRelativePrice = [];
+      allTimeHighPrice = 0;
+      console.log(data.ConversionType.type);
       if (data.ConversionType.type === "divide") {
-        getHistoricalPriceData(coinNumerator, coinDenominator, unixEntered, 1);
+        getHistoricalPriceData(1);
       } else {
         for (var i = 1999; i >= 1; i--) {
           let newDate = new Date((data.Data[i].time + 86400) * 1000).toString();
@@ -77,14 +88,50 @@ function getHistoricalPriceData(
               var dailyChangePrice = dayOfPrice / previousDayPrice - 1;
             }
             historicalChange.push(dailyChangePrice * 100);
+            if (
+              data.Data[i].high > allTimeHighPrice &&
+              data.Data[i].high <= 5 * data.Data[i].close
+            ) {
+              var allTimeHighDate = new Date(
+                (data.Data[i].time + 86400) * 1000
+              );
+              var athRatio = (100 * data.Data[i].close) / allTimeHighPrice;
+              if (athRatio <= 500) {
+                allTimeHighRatio.push(athRatio);
+              } else {
+                allTimeHighRatio.push(100);
+              }
+              allTimeHighRelativePrice.push(data.Data[i].close);
+              var ratioDate = new Date((data.Data[i].time + 86400) * 1000);
+              allTimeHighRatioDates.push(allTimeHighDate);
+              allTimeHighPrice = data.Data[i].high;
+            } else if (allTimeHighPrice !== 0) {
+              var athRatio = (100 * data.Data[i].close) / allTimeHighPrice;
+              if (athRatio <= 500) {
+                allTimeHighRatio.push(athRatio);
+              } else {
+                allTimeHighRatio.push(100);
+              }
+              allTimeHighRelativePrice.push(data.Data[i].close);
+              var ratioDate = new Date((data.Data[i].time + 86400) * 1000);
+              allTimeHighRatioDates.push(ratioDate);
+            }
+            $(".allTimeHighPriceOutput").html(
+              "ATH of " +
+                coinNumerator +
+                ": " +
+                allTimeHighPrice +
+                "</br> on " +
+                allTimeHighDate.toString().substring(4, 15)
+            );
+            var dayOfWeek = newDate.substr(0, 3);
+            historicalWeekday.push(dayOfWeek);
           }
-          var dayOfWeek = newDate.substr(0, 3);
-          historicalWeekday.push(dayOfWeek);
         }
+        getCurrentPriceData(allTimeHighPrice);
         maxDaysHandle(historicalDate);
-        console.log(historicalPrice);
 
-        switch (menuItem) {
+        switch (navMenuPosition) {
           case 0:
             plotPriceHistory(
               historicalDate,
@@ -92,8 +139,8 @@ function getHistoricalPriceData(
               historicalLow,
               historicalHigh,
               historicalOpen,
-              coinDenominator,
-              coinNumerator
+              coinNumerator,
+              coinDenominator
             );
             break;
           case 1:
@@ -106,18 +153,90 @@ function getHistoricalPriceData(
             );
             break;
           case 2:
-            getHistoricalATHData(
+            plotATHHistory(
+              allTimeHighRelativePrice,
+              allTimeHighRatioDates,
+              allTimeHighRatio,
               coinNumerator,
-              coinDenominator,
-              unixEntered,
-              data,
+              coinDenominator
+            );
+            break;
+        }
+      }
+    });
+}
+
+
+function filterHistoricalPriceData() {
+      convert = 0;
+      var historicalPrice = [];
+      var historicalOpen = [];
+      var historicalLow = [];
+      var historicalHigh = [];
+      var historicalDate = [];
+      var historicalChange = [];
+      var historicalWeekday = [];
+      allTimeHighRatio = [];
+      allTimeHighRatioDates = [];
+      allTimeHighRelativePrice = [];
+      allTimeHighPrice = 0;
+      console.log(data.ConversionType.type);
+      if (data.ConversionType.type === "divide") {
+        getHistoricalPriceData(1);
+      } else {
+        for (var i = 1999; i >= 1; i--) {
+          let newDate = new Date((data.Data[i].time + 86400) * 1000).toString();
+          if (data.Data[i].close == []) {
+            i = 0;
+          } else {
+            historicalPrice.push(data.Data[i].close);
+            historicalLow.push(data.Data[i].low);
+            historicalHigh.push(data.Data[i].high);
+            historicalOpen.push(data.Data[i].open);
+            historicalDate.push(convertDateFormat(newDate));
+            var previousDayPrice = data.Data[i - 1].close;
+            var dayOfPrice = data.Data[i].close;
+            if (previousDayPrice === 0) {
+              dailyChangePrice = 0;
+            } else {
+              var dailyChangePrice = dayOfPrice / previousDayPrice - 1;
+            }
+            historicalChange.push(dailyChangePrice * 100);
+            var dayOfWeek = newDate.substr(0, 3);
+            historicalWeekday.push(dayOfWeek);
+          }
+        }
+        getCurrentPriceData(allTimeHighPrice);
+        maxDaysHandle(historicalDate);
+
+        switch (navMenuPosition) {
+          case 0:
+            plotPriceHistory(
+              historicalDate,
               historicalPrice,
               historicalLow,
               historicalHigh,
               historicalOpen,
-              historicalDate,
+              coinNumerator,
+              coinDenominator
+            );
+            break;
+          case 1:
+            plotWeekdayChange(
+              historicalChange,
               historicalWeekday,
-              historicalChange
+              historicalDate,
+              coinNumerator,
+              coinDenominator
+            );
+            break;
+          case 2:
+            plotATHHistory(
+              allTimeHighRelativePrice,
+              allTimeHighRatioDates,
+              allTimeHighRatio,
+              coinNumerator,
+              coinDenominator
             );
             break;
         }
@@ -133,13 +252,11 @@ function getHistoricalPriceData(
 //   historicalDate,
 //   historicalWeekday,
 //   historicalChange,
-//   coinNumerator,
-//   coinDenominator,
 //   allTimeHighRatio,
 //   allTimeHighRatioDates,
 //   allTimeHighRelativePrice
 // ) {
-//   switch (menuItem) {
+//   switch (navMenuPosition) {
 //     case 0:
 //       plotPriceHistory(
 //         historicalDate,
@@ -147,8 +264,6 @@ function getHistoricalPriceData(
 //         historicalLow,
 //         historicalHigh,
 //         historicalOpen,
-//         coinDenominator,
-//         coinNumerator
 //       );
 //       break;
 //     case 1:
@@ -156,75 +271,69 @@ function getHistoricalPriceData(
 //         historicalChange,
 //         historicalWeekday,
 //         historicalDate,
-//         coinNumerator,
-//         coinDenominator
 //       );
 //       break;
 //     case 2:
 //       plotATHHistory(
 //         allTimeHighRelativePrice,
 //         allTimeHighRatioDates,
-//         coinDenominator,
-//         coinNumerator,
 //         allTimeHighRatio
 //       );
 //       break;
 //   }
 // }
 
-$(".submit").click(onSubmit());
+$(".submit").click(onSubmit);
 
 function onSubmit() {
-  var coinNumerator = $("#coinCompare1").val().toUpperCase();
-  var coinDenominator = $("#denom-choices").val().toUpperCase();
-  var monthEntry = $("#dateEntry1").val();
-  var dayEntry = $("#dateEntry2").val();
-  var yearEntry = $("#dateEntry3").val();
-  var enteredDate = new Date(yearEntry, monthEntry - 1, dayEntry);
-  unixEntered = enteredDate.getTime() / 1000 + 24 * 60 * 60;
-  getHistoricalPriceData(coinNumerator, coinDenominator, unixEntered, 0);
-  if (menuItem === 3) {
-    getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered);
+  coinNumerator = $("#coinCompare1").val().toUpperCase();
+  coinDenominator = $("#denom-choices").val().toUpperCase();
+  getHistoricalPriceData(0);
+  if (navMenuPosition === 3) {
+    getHistoricalHourlyData();
   }
 }
 
 $(".nav-price").click(function () {
-  menuItem = 0;
+  navMenuPosition = 0;
   filterDays = 365;
   changeHandle();
   onSubmit();
 });
 
 $(".nav-weekday").click(function () {
-  menuItem = 1;
+  navMenuPosition = 1;
   filterDays = 30;
   changeHandle();
   onSubmit();
 });
 
 $(".nav-ath").click(function () {
-  menuItem = 2;
+  navMenuPosition = 2;
   filterDays = 365;
   changeHandle();
   onSubmit();
 });
 
 $(".nav-hourly").click(function () {
-  menuItem = 3;
+  navMenuPosition = 3;
   filterDays = 30;
   changeHandle();
   onSubmit();
 });
 
-function getCurrentPriceData(coinNumerator, coinDenominator) {
-  $.get(
+function getCurrentPriceData(allTimeHighPrice) {
+  const url =
     "https://min-api.cryptocompare.com/data/price?fsym=" +
-      coinNumerator +
-      "&tsyms=" +
-      coinDenominator +
-      "&e=" +
-      exchange,
-    function (data) {
+    coinNumerator +
+    "&tsyms=" +
+    coinDenominator +
+    "&e=" +
+    exchange;
+
+  fetch(url, options)
+    .then((response) => response.json())
+    .then((data) => {
       currentPrice = data[coinDenominator];
       if (allTimeHighPrice > 0) {
         $(".currentPriceOutput").html(
@@ -249,471 +358,13 @@ function getCurrentPriceData(coinNumerator, coinDenominator) {
             "</br> (100% of ATH)"
         );
       }
-    }
-  );
+    });
 }
 
-function getHistoricalATHData(
-  coinNumerator,
-  coinDenominator,
-  unixEntered,
-  data,
-  historicalPrice,
-  historicalLow,
-  historicalHigh,
-  historicalOpen,
-  historicalDate,
-  historicalWeekday,
-  historicalChange
-) {
-  allTimeHighRatio = [];
-  allTimeHighRatioDates = [];
-  allTimeHighRelativePrice = [];
-  allTimeHighPrice = 0;
-  for (var i = 0; i <= 1999; i++) {
-    if (
-      data.Data[i].high > allTimeHighPrice &&
-      data.Data[i].high <= 5 * data.Data[i].close
-    ) {
-      var allTimeHighDate = new Date((data.Data[i].time + 86400) * 1000);
-      var athRatio = (100 * data.Data[i].close) / allTimeHighPrice;
-      if (athRatio <= 500) {
-        allTimeHighRatio.push(athRatio);
-      } else {
-        allTimeHighRatio.push(100);
-      }
-      allTimeHighRelativePrice.push(data.Data[i].close);
-      var ratioDate = new Date((data.Data[i].time + 86400) * 1000);
-      allTimeHighRatioDates.push(allTimeHighDate);
-      allTimeHighPrice = data.Data[i].high;
-    } else if (allTimeHighPrice !== 0) {
-      var athRatio = (100 * data.Data[i].close) / allTimeHighPrice;
-      if (athRatio <= 500) {
-        allTimeHighRatio.push(athRatio);
-      } else {
-        allTimeHighRatio.push(100);
-      }
-      allTimeHighRelativePrice.push(data.Data[i].close);
-      var ratioDate = new Date((data.Data[i].time + 86400) * 1000);
-      allTimeHighRatioDates.push(ratioDate);
-    }
-  }
-  $(".allTimeHighPriceOutput").html(
-    "ATH of " +
-      coinNumerator +
-      ": " +
-      allTimeHighPrice +
-      "</br> on " +
-      allTimeHighDate.toString().substring(4, 15)
-  );
-  getCurrentPriceData(coinNumerator, coinDenominator);
-
-  setTimeout(function () {
-    plotATHHistory(
-      allTimeHighRelativePrice,
-      allTimeHighRatioDates,
-      coinDenominator,
-      coinNumerator,
-      allTimeHighRatio
-    );
-  }, 1000);
-}
-
-function plotPriceHistory(
-  historicalDate,
-  historicalPrice,
-  historicalLow,
-  historicalHigh,
-  historicalOpen,
-  coinDenominator,
-  coinNumerator
-) {
-  var d3 = Plotly.d3;
-  console.log(historicalPrice);
-  var WIDTH_IN_PERCENT_OF_PARENT = 100,
-    HEIGHT_IN_PERCENT_OF_PARENT = 60;
-
-  var gd3 = d3.select("#myDiv").style({
-    width: WIDTH_IN_PERCENT_OF_PARENT + "%",
-    "margin-left": (100 - WIDTH_IN_PERCENT_OF_PARENT) / 2 + "%",
-
-    height: HEIGHT_IN_PERCENT_OF_PARENT + "vh",
-    "margin-top": (60 - HEIGHT_IN_PERCENT_OF_PARENT) / 2 + "vh",
-    "margin-bottom": 0,
-    "padding-bottom": 0,
-  });
-
-  var gd = gd3.node();
-  var trace3 = {
-    x: historicalDate.slice(0, filterDays),
-    y: historicalPrice.slice(0, filterDays),
-    mode: "lines",
-    name: "Close price",
-    type: "scatter",
-  };
-
-  var trace1 = {
-    x: historicalDate.slice(0, filterDays),
-    close: historicalPrice.slice(0, filterDays),
-    open: historicalOpen.slice(0, filterDays),
-    high: historicalHigh.slice(0, filterDays),
-    low: historicalLow.slice(0, filterDays),
-    mode: "lines",
-    name: "Close price",
-    type: "candlestick",
-  };
-
-  var layout = {
-    title: "Close Price " + coinNumerator + " vs " + coinDenominator,
-    xaxis: {
-      title: "Date",
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-      rangeslider: {
-        visible: false,
-      },
-    },
-    yaxis: {
-      title: "Price" + "(" + coinDenominator + ")",
-      type: "linear",
-      autorange: true,
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-    },
-    // shapes: [
-    //   {
-    //     type: "line",
-    //     xref: "paper",
-    //     x0: 0,
-    //     x1: 1,
-    //     y0: currentPrice,
-    //     y1: currentPrice,
-    //     line: {
-    //       color: "rgb(50, 11, 96)",
-    //       width: 1,
-    //       dash: "dashdot",
-    //     },
-    //   },
-    // ],
-  };
-
-  if (convert === 0) {
-    var plotData = [trace1];
-  } else {
-    plotData = [trace3];
-  }
-
-  Plotly.newPlot(gd, plotData, layout);
-  $(".radio2").addClass("hide");
-  $(".radio1").removeClass("hide");
-  $("#radio1").click(function () {
-    if (layout.yaxis.type === "log") {
-      layout.yaxis.type = "linear";
-    } else {
-      layout.yaxis.type = "log";
-    }
-    Plotly.newPlot(gd, plotData, layout);
-  });
-  window.onresize = function () {
-    Plotly.Plots.resize(gd);
-  };
-}
-
-function plotWeekdayChange(
-  historicalChange,
-  historicalWeekday,
-  historicalDate,
-  coinNumerator,
-  coinDenominator
-) {
-  var d3 = Plotly.d3;
-
-  var WIDTH_IN_PERCENT_OF_PARENT = 100,
-    HEIGHT_IN_PERCENT_OF_PARENT = 60;
-
-  var gd3 = d3.select("#myDiv").style({
-    width: WIDTH_IN_PERCENT_OF_PARENT + "%",
-    "margin-left": (100 - WIDTH_IN_PERCENT_OF_PARENT) / 2 + "%",
-
-    height: HEIGHT_IN_PERCENT_OF_PARENT + "vh",
-    "margin-top": (60 - HEIGHT_IN_PERCENT_OF_PARENT) / 2 + "vh",
-    "margin-bottom": 0,
-    "padding-bottom": 0,
-  });
-
-  var gd = gd3.node();
-  historicalChange = historicalChange.slice(0, filterDays);
-  historicalWeekday = historicalWeekday.slice(0, filterDays);
-  historicalDate = historicalDate.slice(0, filterDays);
-  var trace1 = {
-    x: historicalWeekday,
-    text: historicalDate,
-    y: historicalChange,
-    mode: "markers",
-    type: "scatter",
-  };
-
-  var groups2 = {};
-
-  $.each(historicalWeekday, function (ind, itm) {
-    if (!groups2[itm]) {
-      groups2[itm] = { weekValues: [] };
-    }
-    groups2[itm].weekValues.push(historicalChange[ind]); // sum values belonging to same key
-  });
-  var weeklyChangeAverage = [];
-  var weeklyChangeMedian = [];
-  var weeklyChangeStd = [];
-  var weeklyCI = [];
-  for (var i = 0; i <= 6; i++) {
-    weeklyChangeAverage.push(
-      math.mean(groups2[historicalWeekday[i]].weekValues)
-    );
-    weeklyChangeMedian.push(
-      math.median(groups2[historicalWeekday[i]].weekValues)
-    );
-    weeklyChangeStd.push(math.std(groups2[historicalWeekday[i]].weekValues));
-    weeklyCI.push(
-      1.96 *
-        (weeklyChangeStd[i] /
-          math.sqrt(groups2[historicalWeekday[i]].weekValues.length))
-    );
-  }
-  var errorBars = {
-    x: historicalWeekday,
-    y: weeklyChangeAverage,
-    error_y: {
-      type: "data",
-      array: weeklyCI,
-      visible: true,
-    },
-    mode: "markers",
-    type: "scatter",
-    name: "Mean + CI",
-  };
-
-  var trace2 = {
-    x: historicalWeekday,
-    y: weeklyChangeMedian,
-    mode: "markers",
-    type: "scatter",
-    name: "Median",
-  };
-
-  if (timeOffset >= 0) {
-    var layout = {
-      hovermode: "closest",
-      title:
-        "<b>Day of Week" +
-        " (" +
-        coinNumerator +
-        ")" +
-        "</b> <br> Use " +
-        Number(24 - timeOffset) +
-        ":00" +
-        " of the previous day to make trades",
-      xaxis: {
-        title: "Date",
-        titlefont: {
-          family: "Courier New, monospace",
-          size: 18,
-          color: "#7f7f7f",
-        },
-        fixedrange: true,
-      },
-      yaxis: {
-        title: "% Change" + "(" + coinDenominator + ")",
-        type: "linear",
-        autorange: true,
-        titlefont: {
-          family: "Courier New, monospace",
-          size: 18,
-          color: "#7f7f7f",
-        },
-        fixedrange: true,
-      },
-      shapes: [
-        {
-          type: "line",
-          x0: "Sun",
-          x1: "Sat",
-          y0: 0,
-          y1: 0,
-          line: {
-            color: "rgb(50, 171, 96)",
-            width: 4,
-            dash: "dashdot",
-          },
-        },
-      ],
-    };
-  } else {
-    var layout = {
-      hovermode: "closest",
-      title:
-        "<b>Day of Week</b> <br> Use " +
-        Number(0 - timeOffset) +
-        ":00" +
-        " of the current day to make trades",
-      xaxis: {
-        title: "Date",
-        titlefont: {
-          family: "Courier New, monospace",
-          size: 18,
-          color: "#7f7f7f",
-        },
-        fixedrange: true,
-      },
-      yaxis: {
-        title: "% Change" + "(" + coinDenominator + ")",
-        type: "linear",
-        autorange: true,
-        titlefont: {
-          family: "Courier New, monospace",
-          size: 18,
-          color: "#7f7f7f",
-        },
-        fixedrange: true,
-      },
-      shapes: [
-        {
-          type: "line",
-          x0: "Sun",
-          x1: "Sat",
-          y0: 0,
-          y1: 0,
-          line: {
-            color: "rgb(50, 171, 96)",
-            width: 4,
-            dash: "dashdot",
-          },
-        },
-      ],
-    };
-  }
-
-  var dataSetPlotted = [errorBars, trace2];
-  var plotData = dataSetPlotted;
-
-  Plotly.newPlot(gd, plotData, layout);
-
-  Plotly.relayout(gd, {
-    "xaxis.categoryarray": ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"],
-  });
-  $(".radio1").addClass("hide");
-  $(".radio2").removeClass("hide");
-  $("#radio2").click(function () {
-    if (dataSetPlotted.length === 1) {
-      dataSetPlotted = [errorBars, trace2];
-    } else {
-      dataSetPlotted = [trace1];
-    }
-    var plotData = dataSetPlotted;
-    Plotly.newPlot(gd, plotData, layout);
-  });
-  window.onresize = function () {
-    Plotly.Plots.resize(gd);
-  };
-}
-
-function plotATHHistory(
-  allTimeHighRelativePrice,
-  allTimeHighRatioDates,
-  coinDenominator,
-  coinNumerator,
-  allTimeHighRatio
-) {
-  var d3 = Plotly.d3;
-  var WIDTH_IN_PERCENT_OF_PARENT = 100,
-    HEIGHT_IN_PERCENT_OF_PARENT = 60;
-
-  var gd3 = d3.select("#myDiv").style({
-    width: WIDTH_IN_PERCENT_OF_PARENT + "%",
-    "margin-left": (100 - WIDTH_IN_PERCENT_OF_PARENT) / 2 + "%",
-
-    height: HEIGHT_IN_PERCENT_OF_PARENT + "vh",
-    "margin-top": (60 - HEIGHT_IN_PERCENT_OF_PARENT) / 2 + "vh",
-    "margin-bottom": 0,
-    "padding-bottom": 0,
-  });
-
-  var gd = gd3.node();
-  var trace1 = {
-    x: allTimeHighRatioDates.slice(-filterDays),
-    y: allTimeHighRatio.slice(-filterDays),
-    mode: "lines",
-    text: allTimeHighRelativePrice.slice(-filterDays),
-    type: "scatter",
-  };
-
-  var layout = {
-    title: "ATH Ratio " + coinNumerator + "/" + coinDenominator,
-    xaxis: {
-      title: "Date",
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-    },
-    yaxis: {
-      title: "ATH %" + "(" + coinDenominator + ")",
-      type: "linear",
-      autorange: true,
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-    },
-    shapes: [
-      {
-        type: "line",
-        xref: "paper",
-        x0: 0,
-        x1: 1,
-        y0: 100.0,
-        y1: 100.0,
-        line: {
-          color: "rgb(50, 171, 96)",
-          width: 4,
-          dash: "dashdot",
-        },
-      },
-    ],
-  };
-
-  var plotData = [trace1];
-
-  Plotly.newPlot(gd, plotData, layout);
-  $(".radio2").addClass("hide");
-  $(".radio1").removeClass("hide");
-  $("#radio1").click(function () {
-    if (layout.yaxis.type === "log") {
-      layout.yaxis.type = "linear";
-    } else {
-      layout.yaxis.type = "log";
-    }
-    Plotly.newPlot(gd, plotData, layout, { editable: true });
-  });
-  window.onresize = function () {
-    Plotly.Plots.resize(gd);
-  };
-}
 changeHandle();
 function changeHandle() {
   $element = $('input[type="range"]');
-  Last2000Days();
+  last2000Days();
   $element
     .rangeslider({
       polyfill: false,
@@ -727,50 +378,17 @@ function changeHandle() {
     .on("input", function (e) {
       var $handle = $(".rangeslider__handle", e.target.nextSibling);
       updateHandle($handle[0], filterDays);
-      var updatedNum = Math.ceil(filterDays);
       $(".filter-one").html(result[filterDays] + " - " + result[0]);
       filterDays = this.value;
-      if (menuItem === 3) {
-        var coinNumeratorInt = $("#coinCompare1").val();
-        var coinNumerator = coinNumeratorInt.toUpperCase();
-        var coinDenominatorInt = $("#denom-choices").val();
-        var coinDenominator = coinDenominatorInt.toUpperCase();
-        var monthEntry = $("#dateEntry1").val();
-        var dayEntry = $("#dateEntry2").val();
-        var yearEntry = $("#dateEntry3").val();
-        var enteredDate = new Date(yearEntry, monthEntry - 1, dayEntry);
-        unixEntered = enteredDate.getTime() / 1000 + 24 * 60 * 60;
-        getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered);
+      if (navMenuPosition === 3) {
+        getHistoricalHourlyData();
       } else {
-        onSubmit();
+        // onSubmit();
       }
     });
   var $handle = $(".rangeslider__handle");
   updateHandle($handle[0], filterDays);
-  var updatedNum = Math.ceil(filterDays);
   $(".filter-one").html(result[filterDays] + " - " + result[0]);
-}
-function Last2000Days() {
-  result = [];
-  for (var i = 0; i < 2000; i++) {
-    var d = new Date();
-    d.setDate(d.getDate() - i);
-    result.push(formatDate(d));
-  }
-}
-
-function formatDate(date) {
-  var dd = date.getDate();
-  var mm = date.getMonth() + 1;
-  var yyyy = date.getFullYear();
-  if (dd < 10) {
-    dd = "0" + dd;
-  }
-  if (mm < 10) {
-    mm = "0" + mm;
-  }
-  date = mm + "/" + dd + "/" + yyyy;
-  return date;
 }
 
 function updateHandle(el, val) {
@@ -785,18 +403,9 @@ function maxDaysHandle(historicalDate) {
 $(".dayEntry").on("keydown", function (e) {
   if (e.which == 13) {
     $element.val(this.value).change();
-    onSubmit();
-    if (menuItem === 3) {
-      var coinNumeratorInt = $("#coinCompare1").val();
-      var coinNumerator = coinNumeratorInt.toUpperCase();
-      var coinDenominatorInt = $("#denom-choices").val();
-      var coinDenominator = coinDenominatorInt.toUpperCase();
-      var monthEntry = $("#dateEntry1").val();
-      var dayEntry = $("#dateEntry2").val();
-      var yearEntry = $("#dateEntry3").val();
-      var enteredDate = new Date(yearEntry, monthEntry - 1, dayEntry);
-      unixEntered = enteredDate.getTime() / 1000 + 24 * 60 * 60;
-      getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered);
+    // onSubmit();
+    if (navMenuPosition === 3) {
+      getHistoricalHourlyData();
     }
     changeHandle();
   }
@@ -804,32 +413,18 @@ $(".dayEntry").on("keydown", function (e) {
 
 $(".coinEntry").on("keydown", function (e) {
   if (e.which == 13) {
-    var coinNumeratorInt = $("#coinCompare1").val();
-    var coinNumerator = coinNumeratorInt.toUpperCase();
-    var coinDenominatorInt = $("#denom-choices").val();
-    var coinDenominator = coinDenominatorInt.toUpperCase();
-    var monthEntry = $("#dateEntry1").val();
-    var dayEntry = $("#dateEntry2").val();
-    var yearEntry = $("#dateEntry3").val();
-    var enteredDate = new Date(yearEntry, monthEntry - 1, dayEntry);
-    var unixEntered = enteredDate.getTime() / 1000 + 24 * 60 * 60;
-    getHistoricalPriceData(coinNumerator, coinDenominator, unixEntered);
-    if (menuItem === 3) {
-      var coinNumeratorInt = $("#coinCompare1").val();
-      var coinNumerator = coinNumeratorInt.toUpperCase();
-      var coinDenominatorInt = $("#denom-choices").val();
-      var coinDenominator = coinDenominatorInt.toUpperCase();
-      var monthEntry = $("#dateEntry1").val();
-      var dayEntry = $("#dateEntry2").val();
-      var yearEntry = $("#dateEntry3").val();
-      var enteredDate = new Date(yearEntry, monthEntry - 1, dayEntry);
-      unixEntered = enteredDate.getTime() / 1000 + 24 * 60 * 60;
-      getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered);
+    coinNumerator = $("#coinCompare1").val().toUpperCase();
+    console.log(coinNumerator);
+    coinDenominator = $("#denom-choices").val().toUpperCase();
+    console.log(coinDenominator);
+    getHistoricalPriceData();
+    if (navMenuPosition === 3) {
+      getHistoricalHourlyData();
     }
   }
 });
 
-function getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered) {
+function getHistoricalHourlyData() {
   var hourlyDate = [];
   var hourlyLongDate = [];
   var hourlyHour = [];
@@ -844,7 +439,7 @@ function getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered) {
       exchange,
     function (data, status) {
       if (data.ConversionType.type === "divide") {
-        getHistoricalHourlyDataAlt(coinNumerator, coinDenominator, unixEntered);
+        getHistoricalHourlyDataAlt();
       } else {
         for (i = 2001 - filterDays * 24; i <= data.Data.length - 1; i++) {
           var hourlyDateInt = new Date(data.Data[i].time * 1000);
@@ -897,20 +492,14 @@ function getHistoricalHourlyData(coinNumerator, coinDenominator, unixEntered) {
           hourlyHour,
           hourlyPrice,
           errorBars,
-          hourlyPriceChange,
-          coinNumerator,
-          coinDenominator
+          hourlyPriceChange
         );
       }
     }
   );
 }
 
-function getHistoricalHourlyDataAlt(
-  coinNumerator,
-  coinDenominator,
-  unixEntered
-) {
+function getHistoricalHourlyDataAlt() {
   var hourlyDate = [];
   var hourlyLongDate = [];
   var hourlyHour = [];
@@ -974,75 +563,10 @@ function getHistoricalHourlyDataAlt(
         hourlyPrice,
         errorBars,
         hourlyPriceChange,
-        coinNumerator,
-        coinDenominator,
         hourlyChangeMedian
       );
     }
   );
-}
-
-function plotHourlyData(
-  hourlyLongDate,
-  hourlyDate,
-  hourlyHour,
-  hourlyPrice,
-  errorBars,
-  hourlyPriceChange,
-  coinNumerator,
-  coinDenominator,
-  hourlyChangeMedian
-) {
-  var trace1 = {
-    x: hourlyHour.slice(-filterDays * 24),
-    y: hourlyPriceChange.slice(-filterDays * 24),
-    mode: "markers",
-    name: "Close price",
-    type: "scatter",
-  };
-
-  var layout = {
-    title: "Hourly Change " + coinNumerator + " vs " + coinDenominator,
-    xaxis: {
-      title: "Hour (Local Time)",
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-    },
-    yaxis: {
-      title: "% Change" + " " + coinDenominator,
-      type: "linear",
-      autorange: true,
-      titlefont: {
-        family: "Courier New, monospace",
-        size: 18,
-        color: "#7f7f7f",
-      },
-      fixedrange: true,
-    },
-  };
-  var dataSetPlotted = errorBars;
-  var plotData = [dataSetPlotted];
-
-  Plotly.newPlot("myDiv", plotData, layout);
-
-  Plotly.relayout("myDiv", {
-    "xaxis.categoryarray": ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"],
-  });
-  $(".radio1").addClass("hide");
-  $(".radio2").removeClass("hide");
-  $("#radio2").click(function () {
-    if (dataSetPlotted === errorBars) {
-      dataSetPlotted = trace1;
-    } else {
-      dataSetPlotted = errorBars;
-    }
-    var plotData = [dataSetPlotted];
-    Plotly.newPlot("myDiv", plotData, layout);
-  });
 }
 
 document
@@ -1072,7 +596,7 @@ jQuery(document).ready(function ($) {
   $(".popup-trigger").on("click", function (event) {
     event.preventDefault();
     $(".popup").addClass("is-visible");
-    $(".popup-container").html(popupInfo[menuItem]);
+    $(".popup-container").html(popupInfo[navMenuPosition]);
   });
 
   //close popup
